added the main code to EloAlgorithm that takes the names and score of a game and updates the players mmr.

+++ Tanca/TANCA/Assets/Player.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Player
 { 
     public string playerName;
-    public int mmr = 1000;
+    public float mmr = 1000;
 
     public Player(string _playerName)
     {
